+++ mycodo/databases/models/user.py
@@ -22,6 +22,9 @@ class User(UserMixin, CRUDMixin, db.Model):
     theme = db.Column(db.VARCHAR(64))
     landing_page = db.Column(db.Text, default='live')
     language = db.Column(db.Text, default=None)  # Force the web interface to use a specific language
+    password_reset_code = db.Column(db.Text, default=None)
+    password_reset_code_expiration = db.Column(db.DateTime, default=None)
+    password_reset_last_request = db.Column(db.DateTime, default=None)
 
     # roles = db.relationship("Role", back_populates="user")